+++ pages/destinations/[destination].jsx
@@ -5,12 +5,12 @@ import PageHead from '../../components/PageHead.jsx'
 import EntryDetails from '../../components/destinations/EntryDetails.jsx'
 import SearchBar from '../../components/SearchBar.jsx'
 import VisaRequirementMessage from '../../components/destinations/VisaRequirementMessage.jsx'
-import VisaInfoCard from '../../components/destinations/VisaInfoCard.jsx'
-import TravelBan from '../../components/TravelBan.jsx'
-import Schengen from '../../components/Schengen.jsx'
-import KETA from '../../components/KETA.jsx'
-import GCC from '../../components/GCC.jsx'
-import IraqiKurdistanRegion from '../../components/IraqiKurdistanRegion.jsx'
+import VisaInfoCard from '../../components/destinations/infocards/VisaInfoCard.jsx'
+import TravelBan from '../../components/destinations/infocards/TravelBan.jsx'
+import Schengen from '../../components/destinations/infocards/Schengen.jsx'
+import KETA from '../../components/destinations/infocards/KETA.jsx'
+import GCC from '../../components/destinations/infocards/GCC.jsx'
+import IraqiKurdistanRegion from '../../components/destinations/infocards/IraqiKurdistanRegion.jsx'
 import ElectronicVisaApplyButton from '../../components/ElectronicVisaApplyButton.jsx'
 import { displayVisaMessage, displayTravelBan, displaySchengenInfo, displayKETA, displayIraqiKurdistan, displayEVisaButton } from '../../components/destinations/VisaRequirementMessage.jsx'
 import { countries, isoCodesList, schengenCountries } from '../../utils/countrydata.js'
